Move Test*TestSuite functions back to ./tests/{xdc,ndc}/*_test.go
