+++ tests/ndc/ndc_test.go
@@ -1,0 +1,13 @@
+package ndc
+
+import (
+	"flag"
+	"testing"
+
+	"github.com/stretchr/testify/suite"
+)
+
+func TestNDCFuncTestSuite(t *testing.T) {
+	flag.Parse()
+	suite.Run(t, new(NDCFunctionalTestSuite))
+}
