+++ tests/ndc/replication_migration_back_test.go
@@ -1,0 +1,14 @@
+package ndc
+
+import (
+	"flag"
+	"testing"
+
+	"github.com/stretchr/testify/suite"
+)
+
+func TestReplicationMigrationBackTest(t *testing.T) {
+	flag.Parse()
+	suite.Run(t, new(ReplicationMigrationBackTestSuite))
+
+}
