+++ tests/ndc/replication_task_batching_test.go
@@ -1,0 +1,13 @@
+package ndc
+
+import (
+	"flag"
+	"testing"
+
+	"github.com/stretchr/testify/suite"
+)
+
+func TestNDCReplicationTaskBatching(t *testing.T) {
+	flag.Parse()
+	suite.Run(t, new(NDCReplicationTaskBatchingTestSuite))
+}
