+++ tests/xdc/advanced_visibility_test.go
@@ -1,0 +1,13 @@
+package xdc
+
+import (
+	"flag"
+	"testing"
+
+	"github.com/stretchr/testify/suite"
+)
+
+func TestAdvVisCrossDCTestSuite(t *testing.T) {
+	flag.Parse()
+	suite.Run(t, new(AdvVisCrossDCTestSuite))
+}
