+++ tests/xdc/stream_based_replication_test.go
@@ -1,0 +1,11 @@
+package xdc
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/suite"
+)
+
+func TestStreamBasedReplicationTestSuite(t *testing.T) {
+	suite.Run(t, new(streamBasedReplicationTestSuite))
+}
