+++ tests/xdc/user_data_replication_test.go
@@ -1,0 +1,13 @@
+package xdc
+
+import (
+	"flag"
+	"testing"
+
+	"github.com/stretchr/testify/suite"
+)
+
+func TestUserDataReplicationTestSuite(t *testing.T) {
+	flag.Parse()
+	suite.Run(t, new(UserDataReplicationTestSuite))
+}
